Update existing intent rather than replace it
For the use case of using the notification as a timer. If you update the
notification frequently, then tapping it would typically not resume the
app as the PendingIntent had been cancelled. Instead updates to the
notification now update the PendingIntent rather than replacing it, which
makes sense for most cases.