Change iOS UIBackgroundModes value to "remote-notification"
iOS Developer Library reference:

https://developer.apple.com/library/ios/documentation/General/Reference/InfoPlistKeyReference/Articles/iPhoneOSKeys.html#//apple_ref/doc/uid/TP40009252-SW22

"""The app uses remote notifications as a signal that there is new
content available for download. When a remote notification arrives, the
system launches or resumes the app in the background and gives it a
small amount of time to download the new content."""

This value is supported in iOS 7.0 and later.