Fix ios other apps audio interruption (#209)
* Fix ios other apps audio interruption

Other apps audio was interrupted when the background mode was enabled. I found the solution on this post http://stackoverflow.com/questions/10180500/how-to-use-kaudiosessionproperty-overridecategorymixwithothers
If we don't disable the audio session before that the option AVAudioSessionCategoryOptionMixWithOthers is set, it interrupts the other apps audio. Because the category 'AVAudioSessionCategoryPlayback' is set and by default the session is already active and this category interrupts other apps audio without the option 'mix with others'.

* Update APPBackgroundMode.m